🗃️ Create a MySQL compatibility layer for anyquery
This big commit introduces the possibility for any query to be queried over the MySQL wire protocol.

Because most MySQL clients are doing schema discovery, they expect some information_schema tables and functions to exist. In order to be compatible with these clients, Anyquery features a query rewriter so that these queries can be executed by SQLite.