💄 Create a function to output tables
In order to have a clean way to output data, I've made the struct outputFlag.
It converts a table to CSV, JSON, JSONL, plain (tsv), pretty, etc.
Using the InferFlags method, it can figure out automatically the right output format (check the format flag and if the file is a tty)