🚑️ Fix profile loading with array params
If a profile had an array for the user config, it would crash because encoding/json unmarshals as []interface{},
but gob, which serialize the data between the host and the plugin doesn't support it.
To solve this issue, a custom unmarshaller for the user config now transforms []interface{} to []string or []int or []float or []bool.

By fixing this issue, anyquery has now better type safety for the user config.