✨ Make the JSON output stream
Before, when exporting a query to JSON, anyquery was buffering the result in memory before writing to stdout.

Now, it encodes and writes the JSON on the fly so that you can print larger than memory JSON files. It's also slightly faster (an 8MB JSON takes 2.1s rather than 2.5s)