✨ Strengthen log querying capabilities
Before this commit, when querying logs, if the first 1000 rows didn't match all columns, those columns would not be included in the response.

Now, the Grok pattern is parsed by a fork of vjeantet/grok, and therefore, we know the column names beforehand.

Querying is also a bit faster because the RegExp is not recompiled line by line.

The log querying documentation has also been updated.