➖ Ditch pg_query parser for SQL query
Sometimes, Anyquery has to rewrite SQL queries on the fly.
To do so, it has used a combination of a MySQL parser and a PostgreSQL parser.
This is a bit of a hack, and it's not always reliable.

This commit removes the PostgreSQL parser and uses the MySQL parser (from vitess) for all queries.
The MySQL parsed has been forked (in the other/sqlparser directory) and modified to support more SQL syntax.

This change should make the query rewriting more reliable and easier to maintain.

** Changes **

The SQLparser from vitess.io was moved to the other/sqlparser directory.
I've then modified the yacc grammar to support table functions, and to remove the dual table.

The AST formatter was also modified to support table functions. And regenerated using
astfmtgen github.com/julien040/anyquery/other/sqlparser/...