🔥 Remove statical linking of PRQL
PRQL is a rust library that offers an alternative to SQL.
Because it's written with rust, it causes a lot of issues with statical linking.
This commit removes the statical linking of PRQL and instead uses the prqlc binary to compile the PRQL queries.

Due to this change, users are now required to install the prqlc binary to compile the PRQL queries.
Documentation has therefore been updated to reflect this change.

Also, thanks to this changes, it's now easier to compile anyquery for Linux.
.goreleaser.yaml doesn't use an external VM anymore to compile for Linux
and now uses the zig cc compiler.