🐛 Fix issue parquet quasi-binary import #52
Before, when a column had the BYTE_ARRAY type, anyquery would still try to interpret it at UTF-8, leading to a Replacement Character.

To fix this, Anyquery now make sure a []byte slice is returned rather than a string.

I couldn't find any other way to fix this as parquet-go/parquet-go returns a string even for a parquet BYTE_ARRAY